addressService modified and tested

+++ src/main/java/ua/com/foxminded/university/service/AddressService.java
@@ -1,10 +1,7 @@
 package ua.com.foxminded.university.service;
 
-import static java.util.stream.Collectors.toList;
-
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Stream;
 
 import org.springframework.stereotype.Service;
 
@@ -12,6 +9,8 @@ import ua.com.foxminded.university.dao.AddressDao;
 import ua.com.foxminded.university.dao.StudentDao;
 import ua.com.foxminded.university.dao.TeacherDao;
 import ua.com.foxminded.university.model.Address;
+import ua.com.foxminded.university.model.Student;
+import ua.com.foxminded.university.model.Teacher;
 
 @Service
 public class AddressService {
@@ -26,11 +25,11 @@ public class AddressService {
 	this.studentDao = studentDao;
     }
 
-    public void create(Address address) throws Exception {
+    public void create(Address address) {
 	addressDao.create(address);
     }
 
-    public void update(Address address) throws Exception {
+    public void update(Address address) {
 	addressDao.update(address);
     }
 
@@ -42,34 +41,25 @@ public class AddressService {
 	return addressDao.findById(id);
     }
 
-    public void delete(int id) throws Exception {
-	verifyIdExists(id);
-	verifyAddressIsFree(id);
-	addressDao.delete(id);
+    public void delete(int id) {
+	var canDelete = true;
+	canDelete = canDelete && verifyIdExists(id);
+	canDelete = canDelete && verifyAddressIsFree(id);
+	if (canDelete) {
+	    addressDao.delete(id);
+	} else {
+	    System.out.println("Can't delete address");
+	}
+
     }
 
-    private void verifyIdExists(int id) throws Exception {
-	if (!addressDao.findById(id).isPresent()) {
-	    throw new Exception("Address not found");
-	}
+    private boolean verifyIdExists(int id) {
+	return addressDao.findById(id).isPresent();
     }
 
-    private void verifyAddressIsFree(int id) throws Exception {
-	var address = addressDao.findById(id).get();
-	List<Address> teacherAddresses = teacherDao.findAll()
-		.stream()
-		.flatMap(p -> Stream.of(p.getAddress()))
-		.collect(toList());
-	if (teacherAddresses.contains(address)) {
-	    throw new Exception("Address is used by a teacher account, can't delete address");
-	}
-	List<Address> studentAddresses = studentDao.findAll()
-		.stream()
-		.flatMap(p -> Stream.of(p.getAddress()))
-		.collect(toList());
-	if (studentAddresses.contains(address)) {
-	    throw new Exception("Address is used by a student account, can't delete address");
-	}
+    private boolean verifyAddressIsFree(int id) {
+	Optional<Teacher> teacher = teacherDao.findByAddressId(id);
+	Optional<Student> student = studentDao.findByAddressId(id);
+	return teacher.isEmpty() && student.isEmpty();
     }
-
 }
